refactor : 훅 분리
리펙토링
