+++ src/pages/LandingPage/LandingComponent/BannerImages.jsx
@@ -1,49 +1,16 @@
-import { useState, useEffect, useRef } from "react";
 import { FaChevronLeft, FaChevronRight } from "react-icons/fa";
 import bannerImage1 from "../../../assets/images/notice1.png";
 import bannerImage2 from "../../../assets/images/notice2.png";
 import bannerImage3 from "../../../assets/images/notice3.png";
 import bannerImage4 from "../../../assets/images/notice4.png";
+import useBannerSlider from "../../../features/Landing/useBannerSlider";
 
 const bannerImages = [bannerImage1, bannerImage2, bannerImage3, bannerImage4];
 
 const BannerImages = () => {
-  const [index, setIndex] = useState(0);
-  const [paused, setPaused] = useState(false);
-  const elapsed = useRef(0);
-  const timer = useRef(null);
-
-  const DURATION = 10000;
-  const INTERVAL = 50;
-
-  const resetTimer = () => {
-    clearInterval(timer.current);
-    elapsed.current = 0;
-  };
-
-  const goToPrev = () => {
-    resetTimer();
-    setIndex((prev) => (prev === 0 ? bannerImages.length - 1 : prev - 1));
-  };
-
-  const goToNext = () => {
-    resetTimer();
-    setIndex((prev) => (prev + 1) % bannerImages.length);
-  };
-
-  useEffect(() => {
-    if (paused) return;
-
-    timer.current = setInterval(() => {
-      elapsed.current += INTERVAL;
-
-      if (elapsed.current >= DURATION) {
-        goToNext(); // 다음 배너로 이동하면서 resetTimer 호출됨
-      }
-    }, INTERVAL);
-
-    return () => clearInterval(timer.current);
-  }, [index, paused]);
+  const { index, goToPrev, goToNext, setPaused } = useBannerSlider(
+    bannerImages.length
+  );
 
   return (
     <div
